+++ example/components/custom-component.js
@@ -1,17 +1,17 @@
 import React, { Component } from 'react';
 import { Map, TileLayer, Marker, Popup } from '../../src';
 
-const MyPopupMarker = ({ map, position, children }) => (
-  <Marker map={map} position={position}>
+const MyPopupMarker = ({ layerContainer, map, position, children }) => (
+  <Marker layerContainer={layerContainer} map={map} position={position}>
     <Popup>
       <span>{children}</span>
     </Popup>
   </Marker>
 );
 
-const MyMarkersList = ({ map, markers }) => {
+const MyMarkersList = ({ layerContainer, map, markers }) => {
   const items = markers.map(({ key, ...props }) => (
-      <MyPopupMarker key={key} map={map} {...props} />
+      <MyPopupMarker key={key} layerContainer={layerContainer} map={map} {...props} />
   ));
   return <div style={{display: 'none'}}>{items}</div>;
 };
@@ -32,7 +32,7 @@ export default class CustomComponent extends Component {
     const markers = [
       {key: 'marker1', position: [51.5, -0.1], children: 'My first popup'},
       {key: 'marker2', position: [51.51, -0.1], children: 'My second popup'},
-      {key: 'marker3', position: [51.49, -0.05], children: 'My third popup'}
+      {key: 'marker3', position: [51.49, -0.05], children: 'My third popup'},
     ];
     return (
       <Map center={center} zoom={this.state.zoom}>