+++ src/MapLayer.js
@@ -1,17 +1,17 @@
 import { assign } from 'lodash';
-import React, { PropTypes } from 'react';
+import React, { cloneElement, Children, PropTypes } from 'react';
 import { Map } from 'leaflet';
-import { layerContainer as layerContainerType } from './types';
+
+import childrenType from './types/children';
+import layerContainerType from './types/layerContainer';
+
 import MapComponent from './MapComponent';
 
 export default class MapLayer extends MapComponent {
   static propTypes = {
-    children: PropTypes.oneOfType([
-      PropTypes.arrayOf(PropTypes.node),
-      PropTypes.node,
-    ]),
-    layerContainer: layerContainerType.isRequired,
-    map: PropTypes.instanceOf(Map).isRequired,
+    children: childrenType,
+    layerContainer: layerContainerType,
+    map: PropTypes.instanceOf(Map),
   };
 
   componentDidMount() {
@@ -24,17 +24,17 @@ export default class MapLayer extends MapComponent {
     this.props.layerContainer.removeLayer(this.leafletElement);
   }
 
-  getClonedChildrenWithMap(extra) {
+  getClonedChildrenWithProps(extra) {
     const { children, map, layerContainer } = this.props;
     const props = assign({map, layerContainer}, extra);
 
-    return React.Children.map(children, child => {
-      return child ? React.cloneElement(child, props) : null;
+    return Children.map(children, child => {
+      return child ? cloneElement(child, props) : null;
     });
   }
 
   renderChildrenWithProps(props) {
-    const children = this.getClonedChildrenWithMap(props);
+    const children = this.getClonedChildrenWithProps(props);
     return <div style={{display: 'none'}}>{children}</div>;
   }
 }