+++ src/types/index.js
@@ -1,4 +1,5 @@
 export bounds from './bounds';
+export children from './children';
 export controlPosition from './controlPosition';
 export latlng from './latlng';
 export latlngList from './latlngList';